requirements.txt
torch
torchvision
torchaudio
torchtext
spacy
nltk
numpy 
pandas
matplotlib
librosa
scikit-learn
scipy
transformers
datasets
torchmetrics
tqdm
wandb
sentencepiece
scikit-image
opencv-python
seaborn
plotly
jupyterlab
jupyterlab-git
ipywidgets
ipykernel
huggingface_hub
openai-whisper